+++ R/custom_fit_model.R
@@ -4,6 +4,11 @@ custom_fit_model <- function(model_specifics, train_set){
   model_formula <- model_specifics[["model_formula"]]
   model_fn <- model_specifics[["model_fn"]]
 
+  # Some model types don't accept string formulas
+  # so we convert it to an actual formula object
+  if (is.character(model_formula))
+    model_formula <- as.formula(model_formula)
+
   model_fn(formula = model_formula,
            data = train_set)