+++ R/custom_model_fn.R
@@ -2,10 +2,14 @@
 
 custom_model_fn <- function(train_data,
                             test_data,
-                            fold_info = list(rel_fold=NULL, abs_fold=NULL, fold_column=NULL),
-                            model_specifics = list(model_formula=NULL, family=NULL, link=NULL,
-                                                   control=NULL, REML=FALSE, positive=2,
-                                                   cutoff=0.5, model_verbose=FALSE, model_fn=NULL)){
+                            fold_info = list(rel_fold = NULL,
+                                             abs_fold = NULL,
+                                             fold_column = NULL),
+                            model_specifics = list(
+                              model_formula = NULL, family = NULL, link = NULL,
+                              control = NULL, REML = FALSE, positive = 2,
+                              cutoff = 0.5, model_verbose = FALSE, model_fn = NULL,
+                              caller = NULL)){
 
   # Make sure, a model function was actually passed
   if (is.null(model_specifics[["model_fn"]])){
@@ -31,7 +35,8 @@ custom_model_fn <- function(train_data,
     warn_info = list(
       model_formula = model_specifics[["model_formula"]],
       fold_info = fold_info,
-      model_verbose = model_specifics[["model_verbose"]]
+      model_verbose = model_specifics[["model_verbose"]],
+      caller = model_specifics[["caller"]]
     ))
 
   model <- fitted_model[["model"]]
@@ -41,14 +46,45 @@ custom_model_fn <- function(train_data,
   #   Create a list of NA predictions the length of y_column
 
   if (is.null(model)){
-    predictions <- rep(NA, length(test_data[[y_col]]))
+
+    predictions <- tibble::tibble("prediction" = rep(NA, length(test_data[[y_col]])))
+
   } else {
-    if (model_specifics[["family"]] == "gaussian")
-      predictions <- stats::predict(model, test_data, allow.new.levels=TRUE)
-    if (model_specifics[["family"]] == "binomial")
-      predictions <- stats::predict(model, test_data, type="response", allow.new.levels=TRUE)
-    if (model_specifics[["family"]] == "multinomial")
-      predictions <- stats::predict(model, test_data, type="probs", allow.new.levels=TRUE)
+
+    if (model_specifics[["family"]] == "gaussian"){
+      predictions <- tibble::enframe(
+        stats::predict(model, test_data, allow.new.levels = TRUE),
+        value = "prediction") %>% dplyr::select(.data$prediction)
+
+      # Force type numeric
+      predictions[["prediction"]] <- force_numeric(
+        predictions_vector = predictions[["prediction"]],
+        caller = model_specifics[["caller"]])
+
+      # Select prediction column
+      predictions <- predictions %>%
+        dplyr::select(.data$prediction)
+
+    } else if (model_specifics[["family"]] == "binomial"){
+      predictions <- tibble::enframe(
+        stats::predict(model, test_data, type = "response",
+                       allow.new.levels = TRUE),
+        value = "prediction")
+
+      # Force type numeric and extract prediction column
+      predictions[["prediction"]] <- force_numeric(
+        predictions_vector = predictions[["prediction"]],
+        caller = model_specifics[["caller"]])
+
+      # Select prediction column
+      predictions <- predictions %>%
+        dplyr::select(.data$prediction)
+
+    } else if (model_specifics[["family"]] == "multinomial"){
+      predictions <- dplyr::as_tibble(
+        stats::predict(model, test_data, type = "probs",
+                       allow.new.levels = TRUE))
+    }
   }
 
   predictions_and_targets <- tibble::tibble("target" = test_data[[y_col]]) %>%
@@ -66,4 +102,29 @@ custom_model_fn <- function(train_data,
   )
 }
 
+
+force_numeric <- function(predictions_vector, caller = ""){
+
+  if (!is.numeric(predictions_vector)) {
+    predictions_vector <- tryCatch({
+      if (is.factor(predictions_vector)) {
+        predictions_vector <- as.numeric(as.character(predictions_vector))
+      } else {
+        as.numeric(predictions_vector)
+      }
+    }, error = function(e) {
+      stop(paste0(
+        caller,
+        ": Could not convert predictions to type numeric."
+      ))
+    }, warning = function(w) {
+      warning(paste0(
+        caller,
+        ": Could not convert predictions to type numeric."
+      ))
+    })
+  }
+
+  predictions_vector
+}