+++ tests/testthat/test_cross_validate_fn.R
@@ -1,0 +1,212 @@
+library(cvms)
+context("cross_validate_fn()")
+
+
+test_that("binomial glm models work with cross_validate_fn()",{
+
+  # Load data and fold it
+  set_seed_for_R_compatibility(1)
+  dat <- groupdata2::fold(participant.scores, k = 4,
+                          cat_col = 'diagnosis',
+                          id_col = 'participant')
+
+  glm_model_fn <- function(data, formula){
+    glm(formula = formula, data = data, family = "binomial")
+  }
+
+  CVbinomlist <- cross_validate_fn(dat,
+                                   glm_model_fn,
+                                   formulas = c("diagnosis~score","diagnosis~age"),
+                                   fold_cols = '.folds', type = 'binomial',
+                                   model_verbose = FALSE,
+                                   positive = 1)
+
+  expect_equal(CVbinomlist$AUC, c(0.7615741, 0.1666667), tolerance=1e-3)
+  expect_equal(CVbinomlist$`Lower CI`, c(0.58511535, 0.01748744), tolerance=1e-3)
+  expect_equal(CVbinomlist$`Upper CI`, c(0.9380328, 0.3158459), tolerance=1e-3)
+  expect_equal(CVbinomlist$Kappa, c(0.4927536, -0.3636364), tolerance=1e-3)
+  expect_equal(CVbinomlist$Sensitivity, c(0.5833333,0.0000000), tolerance=1e-3)
+  expect_equal(CVbinomlist$Specificity, c(0.8888889,0.6666667), tolerance=1e-3)
+  expect_equal(CVbinomlist$`Pos Pred Value`, c(0.7777778, 0.0000000), tolerance=1e-3)
+  expect_equal(CVbinomlist$`Neg Pred Value`, c(0.7619048, 0.5), tolerance=1e-3)
+  expect_equal(CVbinomlist$F1, c(0.6666667, NA), tolerance=1e-3)
+  expect_equal(CVbinomlist$Prevalence, c(0.4,0.4), tolerance=1e-3)
+  expect_equal(CVbinomlist$`Detection Rate`, c(0.2333333,0.0000000), tolerance=1e-3)
+  expect_equal(CVbinomlist$`Detection Prevalence`, c(0.3,0.2), tolerance=1e-3)
+  expect_equal(CVbinomlist$`Balanced Accuracy`, c(0.7361111,0.3333333), tolerance=1e-3)
+  expect_equal(CVbinomlist$MCC, c(0.5048268, -0.4082483), tolerance=1e-3)
+  expect_equal(CVbinomlist$Folds, c(4,4))
+  expect_equal(CVbinomlist$`Fold Columns`, c(1,1))
+  expect_equal(CVbinomlist$`Convergence Warnings`, c(0,0))
+  expect_equal(CVbinomlist$Family, c('binomial','binomial'))
+  expect_equal(CVbinomlist$Dependent, c('diagnosis','diagnosis'))
+  expect_equal(CVbinomlist$Fixed, c('score','age'))
+
+  # Enter sub tibbles
+  expect_is(CVbinomlist$Predictions[[1]], "tbl_df")
+  expect_is(CVbinomlist$ROC[[1]], "tbl_df")
+  expect_equal(colnames(CVbinomlist$Predictions[[1]]), c("Fold Column","Fold","Target","Prediction","Predicted Class"))
+  expect_equal(colnames(CVbinomlist$ROC[[1]]), c("Sensitivities","Specificities"))
+  expect_equal(nrow(CVbinomlist$Predictions[[1]]),30)
+  expect_equal(nrow(CVbinomlist$ROC[[1]]),29)
+
+})
+
+
+test_that("gaussian lm model with cross_validate_fn()",{
+
+  # skip_test_if_old_R_version()
+
+  # Load data and fold it
+  set_seed_for_R_compatibility(1)
+  dat <- groupdata2::fold(participant.scores, k = 4,
+                          cat_col = 'diagnosis',
+                          id_col = 'participant')
+
+  lm_model_fn <- function(data, formula){
+    lm(formula = formula, data = data)
+  }
+  # summary(lmm <- lm_model_fn(dat, "score ~ diagnosis"))
+  # MuMIn::AICc(lmm, REML = F) # The one used in the package
+
+  # Cross-validate the data
+  CVed <- cross_validate_fn(dat,
+                            model_fn = lm_model_fn,
+                            formulas = "score~diagnosis",
+                            fold_cols = '.folds',
+                            type = 'gaussian',
+                            model_verbose = FALSE)
+
+  expect_equal(CVed$RMSE, 17.16817, tolerance=1e-3)
+  expect_equal(CVed$MAE, 14.26914, tolerance=1e-3)
+  expect_equal(CVed$r2m, 0.2640793, tolerance=1e-3)
+  expect_equal(CVed$r2c, 0.2640793, tolerance=1e-3)
+  expect_equal(CVed$AIC, 194.6904, tolerance=1e-3)
+  expect_equal(CVed$AICc, 195.9963, tolerance=1e-3)
+  expect_equal(CVed$BIC, 198.0243, tolerance=1e-3)
+  expect_equal(CVed$Folds, 4)
+  expect_equal(CVed$`Fold Columns`, 1)
+  expect_equal(CVed$`Convergence Warnings`, 0)
+  expect_equal(CVed$Family, 'gaussian')
+  expect_equal(CVed$Dependent, 'score')
+  expect_equal(CVed$Fixed, 'diagnosis')
+
+})
+
+test_that("binomial svm models from e1071 work with cross_validate_fn()",{
+
+  # Load data and fold it
+  set_seed_for_R_compatibility(1)
+  dat <- groupdata2::fold(participant.scores, k = 4,
+                          cat_col = 'diagnosis',
+                          id_col = 'participant')
+  dat[["diagnosis"]] <- factor(dat[["diagnosis"]])
+
+  svm_model_fn <- function(data, formula){
+
+    e1071::svm(formula = formula, # converted to formula object within custom_fit_model()
+               data = data,
+               kernel = "linear",
+               cost = 10,
+               scale = FALSE,
+               type = "C-classification")
+  }
+  # sm_ <- svm_model_fn(data = dat, formula = "diagnosis~score")
+
+  CVbinomlist <- cross_validate_fn(dat,
+                                   svm_model_fn,
+                                   formulas = c("diagnosis~score","diagnosis~age"),
+                                   fold_cols = '.folds', type = 'binomial',
+                                   model_verbose = FALSE,
+                                   positive = 1)
+
+  expect_equal(CVbinomlist$AUC, c(0.736111111111111, 0.5), tolerance=1e-3)
+  expect_equal(CVbinomlist$`Lower CI`, c(0.572405153339587, 0.5), tolerance=1e-3)
+  expect_equal(CVbinomlist$`Upper CI`, c(0.899817068882636, 0.5), tolerance=1e-3)
+  expect_equal(CVbinomlist$Kappa, c(0.492753623188406, 0), tolerance=1e-3)
+  expect_equal(CVbinomlist$Sensitivity, c(0.5833333,0.0000000), tolerance=1e-3)
+  expect_equal(CVbinomlist$Specificity, c(0.8888889,1.0), tolerance=1e-3)
+  expect_equal(CVbinomlist$`Pos Pred Value`, c(0.7777778, NaN), tolerance=1e-3)
+  expect_equal(CVbinomlist$`Neg Pred Value`, c(0.7619048, 0.6), tolerance=1e-3)
+  expect_equal(CVbinomlist$F1, c(0.6666667, NA), tolerance=1e-3)
+  expect_equal(CVbinomlist$Prevalence, c(0.4,0.4), tolerance=1e-3)
+  expect_equal(CVbinomlist$`Detection Rate`, c(0.2333333,0.0000000), tolerance=1e-3)
+  expect_equal(CVbinomlist$`Detection Prevalence`, c(0.3,0.0), tolerance=1e-3)
+  expect_equal(CVbinomlist$`Balanced Accuracy`, c(0.7361111,0.5), tolerance=1e-3)
+  expect_equal(CVbinomlist$MCC, c(0.5048268, 0), tolerance=1e-3)
+  expect_equal(CVbinomlist$Folds, c(4,4))
+  expect_equal(CVbinomlist$`Fold Columns`, c(1,1))
+  expect_equal(CVbinomlist$`Convergence Warnings`, c(0,0))
+  expect_equal(CVbinomlist$Family, c('binomial','binomial'))
+  expect_equal(CVbinomlist$Dependent, c('diagnosis','diagnosis'))
+  expect_equal(CVbinomlist$Fixed, c('score','age'))
+
+  # Enter sub tibbles
+  expect_is(CVbinomlist$Predictions[[1]], "tbl_df")
+  expect_is(CVbinomlist$ROC[[1]], "tbl_df")
+  expect_equal(colnames(CVbinomlist$Predictions[[1]]), c("Fold Column","Fold","Target","Prediction","Predicted Class"))
+  expect_equal(colnames(CVbinomlist$ROC[[1]]), c("Sensitivities","Specificities"))
+  expect_equal(nrow(CVbinomlist$Predictions[[1]]),30)
+  expect_equal(nrow(CVbinomlist$ROC[[1]]),3)
+
+})
+
+
+test_that("gaussian svm models from e1071 work with cross_validate_fn()",{
+
+  # skip_test_if_old_R_version()
+
+  # Load data and fold it
+  set_seed_for_R_compatibility(1)
+  dat <- groupdata2::fold(participant.scores, k = 4,
+                          cat_col = 'diagnosis',
+                          id_col = 'participant')
+
+  svm_model_fn <- function(data, formula){
+
+    e1071::svm(formula = formula, # converted to formula object within custom_fit_model()
+               data = data,
+               kernel = "linear",
+               cost = 10,
+               scale = FALSE,
+               type = "eps-regression")
+  }
+
+  # Cross-validate the data
+  CVed <- cross_validate_fn(dat,
+                            model_fn = svm_model_fn,
+                            formulas = "score~diagnosis",
+                            fold_cols = '.folds',
+                            type = 'gaussian',
+                            model_verbose = FALSE)
+
+  expect_equal(CVed$RMSE, 18.01026, tolerance=1e-3)
+  expect_equal(CVed$MAE, 15.27778, tolerance=1e-3)
+  expect_equal(CVed$r2m, NaN, tolerance=1e-3)
+  expect_equal(CVed$r2c, NaN, tolerance=1e-3)
+  expect_equal(CVed$AIC, NaN, tolerance=1e-3)
+  expect_equal(CVed$AICc, NaN, tolerance=1e-3)
+  expect_equal(CVed$BIC, NaN, tolerance=1e-3)
+  expect_equal(CVed$Folds, 4)
+  expect_equal(CVed$`Fold Columns`, 1)
+  expect_equal(CVed$`Convergence Warnings`, 0)
+  expect_equal(CVed$`Singular Fit Messages`, 0)
+  expect_equal(CVed$Family, 'gaussian')
+  expect_equal(CVed$Dependent, 'score')
+  expect_equal(CVed$Fixed, 'diagnosis')
+
+  expect_equal(colnames(CVed$Coefficients[[1]]),
+               c("term", "estimate", "Fold",
+                 "Fold Column", "p.value"))
+  expect_equal(CVed$Coefficients[[1]]$term,
+               rep(c("(Intercept)","diagnosis"), 4))
+  expect_equal(CVed$Coefficients[[1]]$estimate,
+               c(40.1, -10, 50.1, -20, 40.1, -10, 45, -10))
+  expect_equal(CVed$Coefficients[[1]]$Fold,
+               c(1L, 1L, 2L, 2L, 3L, 3L, 4L, 4L))
+  expect_equal(CVed$Coefficients[[1]]$`Fold Column`,
+               rep(".folds", 8))
+  expect_equal(CVed$Coefficients[[1]]$p.value,
+               rep(NA, 8))
+
+})